mysql: fix output handling and connection cleanup
Read @low_stock back after calling the stock_level procedure; the
transaction's only output was never fetched, so low_stock was
always reported as zero.  Close and clear the MYSQL handle when
mysql_real_connect or the autocommit setup fails; the caller
retries in a loop and leaked one handle per attempt.  Check the
mysql_init result before using it.  Supply the missing statement
argument in the payment error log, which printed the literal word
stmt.  Declare the payment procedure's in_h_amount as
NUMERIC(6, 2) instead of INT, which silently truncated the cents
the client sends.

+++ storedproc/mysql/payment.sql
@@ -14,7 +14,7 @@ drop procedure if exists payment;
 delimiter |
 
 CREATE PROCEDURE payment(in_w_id INT, in_d_id INT, in_c_id INT, in_c_w_id INT, in_c_d_id INT,
-                         in_c_last VARCHAR(16), in_h_amount INT)
+                         in_c_last VARCHAR(16), in_h_amount NUMERIC(6, 2))
 BEGIN
 
 DECLARE  out_w_name VARCHAR(10);
